Adding coordinate object for data storage and map.
Along with the location_address object, I have attempted to add the coordinates with their longitude and latitudes. 
My goal was to form an object, location_coords, that held a tuple of the lat and long.
Surely I have done something correctly as my knowledge with Java is limited. 
Let me know if you have a question about what/how I was trying to set up in the chat.
Other than, that I like the beginning that we have so far!